Add dev command Move db init to dev command
