+++ backend/infrahub/cli/__init__.py
@@ -7,9 +7,9 @@ from infrahub.core.initialization import initialization
 from ..workers.dependencies import get_database
 from .context import CliContext
 from .db import app as db_app
+from .dev import app as dev_app
 from .events import app as events_app
 from .git_agent import app as git_app
-from .schema import app as schema_app
 from .server import app as server_app
 from .tasks import app as tasks_app
 from .upgrade import upgrade_cmd
@@ -28,7 +28,7 @@ app.add_typer(git_app, name="git-agent", hidden=True)
 app.add_typer(db_app, name="db")
 app.add_typer(events_app, name="events", help="Interact with the events system.", hidden=True)
 app.add_typer(tasks_app, name="tasks", hidden=True)
-app.add_typer(schema_app, name="schema")
+app.add_typer(dev_app, name="dev", help="Internal development commands.")
 app.command(name="upgrade")(upgrade_cmd)
